+++ src/test/java/nzilbb/labbcat/TestLabbcatView.java
@@ -487,7 +487,7 @@ public class TestLabbcatView {
   @Test(expected = StoreException.class) public void searchInvalidPattern()
     throws Exception {
     String threadId = labbcat.search(
-      Json.createObjectBuilder().build(), null, null, false, false, null, null);
+      Json.createObjectBuilder().build(), null, null, false, null, null, null);
   }
 
   /** Ensure searches can be cancelled. */
@@ -501,7 +501,7 @@ public class TestLabbcatView {
                      .add("orthography", Json.createObjectBuilder()
                           .add("pattern", ".*")))))
       .build();
-    String threadId = labbcat.search(pattern, null, null, false, false, null, null);
+    String threadId = labbcat.search(pattern, null, null, false, null, null, null);
     labbcat.cancelTask(threadId);
   }
 
@@ -516,7 +516,7 @@ public class TestLabbcatView {
 
     // all instances of "and"
     JsonObject pattern = new PatternBuilder().addMatchLayer("orthography", "and").build();
-    String threadId = labbcat.search(pattern, participantId, null, false, false, null, null);
+    String threadId = labbcat.search(pattern, participantId, null, false, null, null, null);
     try {
       TaskStatus task = labbcat.waitForTask(threadId, 30);
       // if the task is still running, it's taking too long, so cancel it
@@ -561,9 +561,9 @@ public class TestLabbcatView {
     // all instances of "mmm", which are frequently used in overlapping speech
     JsonObject pattern = new PatternBuilder().addMatchLayer("orthography", "mmm").build();
     Match[] includingOverlapping = labbcat.getMatches(
-      pattern, null, null, false, false, null, null, 0);
+      pattern, null, null, false, null, null, null, 0);
     Match[] excludingOverlapping = labbcat.getMatches(
-      pattern, null, null, false, false, null, 5, 0);
+      pattern, null, null, false, null, null, 5, 0);
     assertTrue("There are fewer matches when overlapping speech is excluded",
                includingOverlapping.length > excludingOverlapping.length);
   }
@@ -578,7 +578,7 @@ public class TestLabbcatView {
 
     // all instances of "and"
     JsonObject pattern = new PatternBuilder().addMatchLayer("orthography", "and").build();
-    String threadId = labbcat.search(pattern, participantId, null, false, false, null, null);
+    String threadId = labbcat.search(pattern, participantId, null, false, null, null, null);
     try {
       TaskStatus task = labbcat.waitForTask(threadId, 30);
       // if the task is still running, it's taking too long, so cancel it
@@ -625,7 +625,7 @@ public class TestLabbcatView {
 
     // all instances of "and"
     JsonObject pattern = new PatternBuilder().addMatchLayer("orthography", "and").build();
-    String threadId = labbcat.search(pattern, participantId, null, false, false, null, null);
+    String threadId = labbcat.search(pattern, participantId, null, false, null, null, null);
     try {
       TaskStatus task = labbcat.waitForTask(threadId, 30);
       // if the task is still running, it's taking too long, so cancel it